added last functionality todo is just few bug fixes and doc

+++ FiletrackAPI/Services/AzureBlobService.cs
@@ -13,18 +13,20 @@ public interface IAzureBlobService
     Task DeleteBlob(string fileName);
     Task<string> UpdateBlob(string existingFileName, string newFileName);
     Task<Stream> DownloadBlob(string filename);
-    string GenerateBlobFileName(List<PathMember> path, List<JobAttribute> attributes, string FileName);
+    string GenerateBlobFileName(List<PathMember> path, List<JobAttribute> attributes, string fileName);
 }
 
 public class AzureBlobService : IAzureBlobService
 {
     private readonly string _connectionString;
     private readonly string _containerName;
+    private readonly ILogger<AzureBlobService> _logger;
 
-    public AzureBlobService(IOptions<AppSettings> appSettings)
+    public AzureBlobService(IOptions<AppSettings> appSettings,ILogger<AzureBlobService> logger)
     {
         _connectionString = appSettings.Value.BlobConnectionString;
         _containerName = appSettings.Value.BlobContainer;
+        _logger = logger;
     }
 
     public async Task<string> UploadBlob(string fileName,
@@ -32,7 +34,7 @@ public class AzureBlobService : IAzureBlobService
     {
         var blobClient = new BlobContainerClient(_connectionString, _containerName);
         var blob = blobClient.GetBlobClient(fileName);
-        if (!blob.Exists())
+        if (!await blob.ExistsAsync())
             await blob.UploadAsync(fileStream);
         return blob.Uri.AbsoluteUri;
     }
@@ -50,11 +52,16 @@ public class AzureBlobService : IAzureBlobService
         var blobContainerClient = new BlobContainerClient(_connectionString, _containerName);
         var existingBlobClient = blobContainerClient.GetBlobClient(existingFileName);
         var newBlobClient = blobContainerClient.GetBlobClient(newFileName);
-        var poller = await newBlobClient.StartCopyFromUriAsync(existingBlobClient.Uri);
-        await poller.WaitForCompletionAsync();
-        await existingBlobClient.DeleteIfExistsAsync();
+        if (await newBlobClient.ExistsAsync())
+        {
+            var poller = await newBlobClient.StartCopyFromUriAsync(existingBlobClient.Uri);
+            await poller.WaitForCompletionAsync();
+            await existingBlobClient.DeleteIfExistsAsync();
 
-        return newBlobClient.Uri.AbsoluteUri;
+            return newBlobClient.Uri.AbsoluteUri;
+        }
+
+        return existingBlobClient.Uri.AbsoluteUri;
     }
 
     public async Task<Stream> DownloadBlob(string filename)
@@ -67,18 +74,18 @@ public class AzureBlobService : IAzureBlobService
         return output;
     }
     
-    public string GenerateBlobFileName(List<PathMember> path, List<JobAttribute> attributes, string FileName)
+    public string GenerateBlobFileName(List<PathMember> path, List<JobAttribute> attributes, string fileName)
     {
         string blobFileName = "";
         for (int i = 0; i < path.Count; i++)
         {
             var member = path.FirstOrDefault(x => x.Order == i);
             var value = attributes.FirstOrDefault(x => x.id == member?.Id);
-            blobFileName += value.value;
+            blobFileName += value?.value;
             blobFileName += "/";
         }
 
-        blobFileName += FileName;
+        blobFileName += fileName;
         return blobFileName;
     }
 }
